+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v1.0.1
 ## Fixes
 * macOS: Added warning (and solution) if `Gupax/P2Pool/XMRig` were quarantined by [`Gatekeeper`](https://support.apple.com/en-us/HT202491)
+* P2Pool/XMRig: Added a red `Start` button on errors (bad PATH, invalid file, etc) and a solution in the tooltip
 * P2Pool: Fixed custom node selection sometimes using old values after save
 * Miscellaneous UI changes